Return the last new message on cache corruption
If the message stack is incompatible with the previous stack, the cache
is considered corrupt and is replaced. This leaves a few possibly-unread
messages sent to the void. To mitigate this at least partially, the last
message of the stack will be compared against the old last message of
the stack. If they don't match, we can safely assume that there's at
least one new message, so we can return that.